requirements.txt
pandas>=2.2
Pillow>=10.0
torch>=2.2
torchvision>=0.17
